AtaraxiaSDK V0.01
This is the first step into creating a launchable window on Mac. We use cocoa as the basis of generating a window.

The first build is a very barebones one, as the simple task of windowing through any library often requires dynamically linking that is hard to achieve on Mac