AtaraxiaSDK V0.03
This was updated to support the introduction of color to our window. Now, when you run this, we get a solid Red, indicating that our SDK is supporting the graphical display of colors.

We are still far from anything exciting happening, as the next task is to get a button working that allows us to cycle through colors