AtaraxiaSDK V0.07
This is some of the updated foundational logic for screen switching. So far, I have yet to figure out exactly how we will get this to work. But I will be testing out a couple of ideas later that should fix the problem.

I have most likely forgot to initialize something, whether that be a function call, or a missing header that doesn't allow me to access said info.

This project is slowly is coming together. There just are a LOT of different things I need to handle in something like this.