AtaraxiaSDK V0.07
There are still some fundamental problems in the windowing aspect.

The logic is getting better and better, I just am still not seeing any return on my efforts.

So trying to figure out exactly what is causing this problem is like finding a needle in the haystack, because there are just so many different working parts with this code, it's a major frustration trying to get things to work.

I'm getting no NSLog hits when I run this code, which will be worth investigating more. Hopefully I'll have an answer by tonight, because this is driving me CRAZY!!!!!