AtaraxiaSDK V0.07
After attempts at fixing this problem, I've deicded to be a bit more careful if how I am using C.

This is a pain in the ass to fix, as window switching has been the bane of my existence. I will be trying some additional steps to ensure that the application works.

Currently, everything is broken, so I'm going to add a WindowDelegate section in hopes that this can fix some of my problems with the mousedown functions.