AtaraxiaSDK V0.08
For the amount of problems I was getting trying to pass error messages, and the refusal for my project to recognize simple NSLogs, I've decided to be a bit more smart about how I go about constructing this SDK.

I am just keeping it in Obj-C, however, I will be writing a euclidean graphical engine (which is optional) eventually in C, since that will store much of the math needed.

There is still quite a lot of more debugging to finish up with. However, I am able to accurately diagnose things since I am getting prompted with logs, which wasn't happening before.

TL;DR, WE ARE SO FUCKING BACK