AtaraxiaSDK V0.08
This should hopefully handle a lot of the various actions a user can take.

This will be the start to how we handle inputs. As these will have important error messages returned if things aren't happening, so I'm not guessing in how I diagnose problems.

Jjfc Obj-C is one of these languages that is hell to pass through to C. So at this point, I've decided to have C come to Obj-C instead. I can bash my head less against the wall by doing this