AtaraxiaSDK V0.08
We have now the window terminating and giving us a log that we've terminated the window. None of my mouse events are properly calling.

So now, it's a matter of creating some additional files to try and fix the problem. Gotta love where Obj-C goes wrong, but at least I am not struggling with getting every little fucking thing to work in Obj-C without the transitional layer.