AtaraxiaSDK V0.08
We still are having issues with the mouse events being run, but luckily, everything else works, as I've doubled checked these with logs.

I will be adding a mouseDelegate to handle mouse events. Beccause there is clearly something wrong with these events being called. There's no problem with my window being called and being initialized, which is why it's best if I make this a seperate piece of code