AtaraxiaSDK V0.08
We have now mouse events recognized as happening. Meaning that my logic for clicking inside a window is correct.

The issue is now that there is a failure to switch to a new screen. It shouldn't require much tweaking left. And I'm glad we are close to where events are being handled correctly.