AtaraxiaSDK V0.08
This was HUGE, as the initial error checking  and approach to color has given us an import of a color sequence. So far, this is huge, as what is being called hits green instead of red.

Jfc this is like the worst debug journey, but we are getting closer and closer