AtaraxiaSDK V0.09
I am working on packaging the application instead for testing, since terminal has a problem of keytesting being impossible without me printing to terminal.

This is the start of the packaged version of this app, so that when I'm diagnosing problems, I have standalone versions of the app to test.