AtaraxiaSDK V0.09
One of the issues I ran into via VSCode is that when I attempt to enter any action, because we were terminal bound, it would print my action to the shell instead of working inside of said window.

So to make this work better for myself when testing. I've decided to go the approach of using a plist to package this into an application that launches itself.

This is part of the program wrapper that enables an app to be built into an executable that is launched.