AtaraxiaSDK V0.09
This makefile builds to an application instead of the target approach. We have to manually reference the plist file by echoing what's contained in it. Then, instead of using target for run, we use open instead.

This is so much less annoying to deal with. Sure, while I may have to create a seperate window and write some additional logic, now, any action I do isn't writing to terminal