AtaraxiaSDK V0.10
I've updated my main file to better integrate my logwindow logic.

The only issue is that the spacebar is not returned as existing or valid, which is a problem. When I fix that, we will increment the software by a version, and tackle the main issue, which is the program properly initializing screen transitions