AtaraxiaSDK V0.10
After a fever dream of trying to rework things, only for me to deliciously fail.

The code is back to where it started, albeit, better organized.

What I will be doing is work on event handling and adding this to the list of logs. Because I am about to split up these files into a million different components, to see exactly what doesn't click