AtaraxiaSDK V0.12
We have added in new logic that enables us to quit the application compared to just closing the window.

We were in a previous state where if we exited the application this way, we never truly terminated the program.

Even though I still face massive issues with the SDK, we are at least moving toward some positive direction.

I will do additional debugging in the morning to try and figure out where these problems lie.

We have to still tackle:
- global event listening
- screen transitions

These direct problems will be handled better with more sleeps. So I snooze in hopes of a brighter day