AtaraxiaSDK V0.13
My attempts to work with audio files in my makefile is far from complete. I am keeping it this way so I don't ruin my entire code.

Yes, I managed to delete everything via not being careful with rm rf when testing things. So, I need to be extra cautious with this section