AtaraxiaSDK V0.14
m4a audio support added.

Audio support is coming along well. After I finish the basics, I'll be taking a break and doing a crosslanguage test with a makefile after.

The frustrations I've faced trying to get these files to work is just a lot of headache. So, I want to start simple so that when I do merge the two languages, I am not getting fundamental problems with the files themselves.

One thing I do want to try, is the ability for my program to work via Windows. If I can get this feat done, it'll allow people to contribute to this code if neccessary.