AtaraxiaSDK V0.15
Multiple makefiles in this build was rather important. Since I need to test parts of my application for the sake of merging code. We want to have a million different ways we can test new ideas of code through makefiles. This is so that when I merge Obj-C++ and C++, I can isolate specific issues, so that when I go to incorporate in these languages into my overall makefile, I do not get errors with make.

This is a fundamental part to the next version planned, which'll be more testing and diagnosing. The command that we processed into our master makefile that worked was:
make -f AtaraxiaSDK/Makefile run

Because I can just pass terminal commands into my makefile, this is a far better approach, albeit, insane, to get things to work. It's rather important because all the other attempts at making this work failed. Really, the best course of action is to prompt your Makefile to return terminal commands instead, since I can verify that each individual makefile worked!

No need to attempt to rewrite the logic of my AtaraxiaSDK makefile into my master makefile, that is where you get a shitton of issues