AtaraxiaSDK V0.15
Now, we can run another part of the application in addition to the current program.

Although I want the two isolated (so we can run them separate for testing), I will take it for now. This project is such a pain with getting makefiles to work, that I'd much rather this than nothing working