AtaraxiaSDK V0.16.1
To illustrate the power of an SDK application, I will be doing a simple FileConverter window that allows the user to convert audio files to valid SDK ones.

This will be a tough integration, as there is a lot more that needs to happen to make this possible. We will be using this as a basis of merging Obj-C++ data with Obj-C data. Since this integration will require a LOT to make this happen. This simple File Converter is gonna be a difficult feat to complete