AtaraxiaSDK V0.16.1
We updated the AppDelegate to support another aspect of the top windowing element, which is a file section that allows us to showcase how capable the SDK is. Since I am getting issues with windowing and global event issues, I should be able to tackle some of the problems I was having here.

This will be the largest project within this project, since we will use this to help fix one of the biggest problems presented in my code. It's exciting that this is happening.

I am just going for general audio support for now, but it would be cool to have a free file converter that works on Mac. I don't want this SDK to be limited to just audio conversion, but common files people use.