AtaraxiaSDK Windows Support V0.1
This absolutely will not run, not in the slightest. I decided to write this .bat file for when I do want to work on support for Windows.

This is not a priority yet, as trying to write an additional translation layer is just too much work for currently what I am attempting. As we get further along in this application, I will add this translation layer, I just know it's gonna be about 500 layers of a headache.

My entire Cocoa library is not friendly for windows, so attempting to get this to run will require Qt. Please do not hope I am primarily targeting Windows, as this is not the main work I am planning on doing. Think of it more as the cherry on top