AtaraxiaSDK V0.16.4
With all the options we have available within the FileConverter window, all that there is left to write is the File Conversion logic. We have all the needed buttons for this to work by now, meaning that I am far closer toward a basic FileConverter.

Audio support will be coming later after I handle basic files that need no Obj-C++. Hopefully the logic part of the vanilla files isn't too difficult to write