AtaraxiaSDK V0.16.4
I've updated this to support more of the export file logic needed for a file converter.

We still are off on getting files to convert, but we are moving in the right direction. This is where this aspect of the project gets more and more complicated, but I am glad to have made this an integral part of my SDK, since it shows the power of development tools like this.

The reason I've chosen to do less relevant files, is that I want the good foundational logic before attempting to proceed with the more complicated parts of the code. Since I haven't even touched the obj-c++ code merging with obj-c, I really don't want to attempt to try something unless I am getting successful file conversions.