AtaraxiaSDK V0.16.4
There are a few more libraries that are essential for this SDK to work, or at least for the code I am implementing.

I purposefully left out the other files from SRC, since I cannot reasonably test these, and the best decisions are left to when I've woken up and have coffee.

So far, this aspect of the code has come a long way, given that I'm getting closer to having file conversion work. It still is a LOT of fucking work, and I have a reasonable fear about just how bad these next few builds are gonna turn out, given these are new libraries, and I hate adding too many new things that might not fuck with Cocoa. Let's just hope my fears are squashed in the morning.