WindowingGUI - Toolbar Menu V0.2
I've updated this main file so that it is the underlying logic needed for an About window to be opened.

There is need for timer related events to trigger this to happen, and I also need to update some of the logic related to Obj-C. We don't want Obj-C creating the window, we want SDL to create it for us. This is the ideal way of doing communication between the two, since managing two different languages for windowing gets into problems.