WindowingGUI - Toolbar Menu V0.3
Fleshed out a good menu system for said application that allows the user to add their own button commands. This is far from where i need it to be, but a good system for building out other ideas via forks.

If you want to support this project, I highly recommend starting here for any menu options you feel is good for the underlying SDK