Documentation Menu Items
This is so that if you need to store loads of documentation within your own app, you can do so. It is overkill for what most people will need, I just thought it would be best to have the help section as a guide for what should typically be stored in said windows.