Bool init windowing approach
In the process of working on a game with SDL3, a better more modular approach to working on windows.

This could be later used to generate additional windows within the software, a crucial software feature of DAWs that need the ability to print multiple screens without the need for a scene transition.