SDL3 Text Library Add
Added in the library so we could generate text.

This is a link to the repository you must clone and build, or else you'll be unable to generate said files:
https://github.com/libsdl-org/SDL_ttf

This is not standardly included in SDL, and there is no homebrew way of installing this library, you will have to clone it, cd into it, and build it. It is the only time I insist on sideloading anything this way, since homebrew is better 99% of the time