SQLite Update - Database for Score Data
For simple games, a Database to keep track of who wins overall is important. This is a very barebones approach, as it only keeps track of a winner or loser.

One aspect I want to incorporate into this SDK is accessing color data from a database file. I may need to play around with how SQL interacts with SDL3, since I do not want this to be a near impossible method of calling to a custom color list and parsing through the data.

I am sticking to SQLite for features like this, since full SQL is just too overboard for a task like this.