SDL3_ttf Integration
Fuck, this was one of the most annoying things on the fucking planet to get done. First, it wouldn't recognize my file location, next, I was having issues of getting the text to initialize after I finished loading it and got the message it was initialized.

TTF_Init() is something you need to make use of anytime you need to load text. This can be loaded into its own function as well and called when needed. I may refactor it into its own functions so it reads better, so that it works more modularly at some point.

SDL3_ttf has been successfully implemented, meaning you can add text to your code and have it work out!