Video Rendering (No Audio)
In the process of loading 'Never Gonna Give You Up', I have written the video implementation behind .mp4 files. This is only for 30 fps or less files for now, as I cannot verify how 60fps video files are handled.

If you do decide to utilize this, make sure that your frame is big enough to handle the video component, or you'll find issues with playback due to the video file being out of bounds with the window.