Video and Audio Handling
This allows you to call to video and audio as needed. Since audio in FFMPEG is a pain, and I couldn't solve it, I've decided to use SDL to handle the audio portion of video.

This is less than ideal, but for now, this is the solution for video handling.

It should tackle audio event triggering, since you'll be able to load and use audio files