Audio Handling V1
This allows you to call to SFX in your project. There is also some cleanup I do within the playSFX function. This is important so that if you switch screens, you don't cause any audio frame mismatch (there was an issue where not clearing this was messing up how audio was interpreted for video calls)