Compliant app bundle render text
One of the biggest parts of my code that needed updating was how I was initializing text. Before, I had this problem that I was constantly initializing text. This was creating a problem during launch of the application, where this constant initialization was eventually making it impossible to locate the font file.

This is fixed, and now you can create an application using make bundle