Makefile
I removed cross platform compatibility in the Makefile, and that had more to do with me diagnosing what my problem was regarding building the application. It was more important for me to make sure I could build the Mac OS app, which I was successful with.

For now, I am only sticking to Mac development for the sake of keeping things simple enough. Eventually when I build things out for Linux, I will be doing a seperate Hello World app build for Linux, same applies with Windows. These are serious projects, and so I will need to ensure that everything I do works while not breaking parts of my code.

SDL has a lot to it, and so the less I focus on trying to do a bunch of difficult things, I usually start with a heavily stripped version.

This update is a serious one, it handles bundling completely for Mac OS X. Meaning you can now genuinely build something that is app standard compliant. This means that this is the first SDK version that was stable enough for me to make an application with it.

Consult my Tic Tac Toe Project since there are some slight version differences in how I'm handling some things.