Modular Update 3
Now we have separate logic in how player is getting rendered that is separate from our main rendering place.

This is key for when we need to render stuff related to particular objects we want to work with, and this modular approach is entirely how we'll approach anything we need independent rendering from our main render logic.

Friendly reminder, if you need additional independent rendering, you need to define it by namespace standards so that you can have multiple functions of the same name