Entity Update
Consolidated things better so that we get proper rendering of entities on screen.

Also wrote the foundation to my tile map entity, which should easily be enough entities you could possible need for a top down scene.

Keep in mind, this needs to be done with scenes. I'll handle the use of scenes sometime soon, I just have refactored enough of this code to where I'm still not there yet to getting to integration with everything else