48x48 Map Tile Support V0.5
This is official support for 48x48 Map support.

However, be warned, if your map tile is a different size compared to mine, it may fall out of the scope of the bounds. This means that you can easily not have enough of a bounds or may have an overextended bounds on what tiles actually exist, meaning you'll get a lot of errors.

At the moment, I have only exposed a reasonable size of tile rendering, the rest exists out in comments for you to utilize as you may see fit.

I'm planning on porting this to a hashmap soon