48x48 Map Tile Support - Official
Instead of substantiating every single tile location which may change depending on what 48x48 map you have imported, we have this in a hashmap. Meaning that you can call to the tile you need, and its position within a map.

This is the first substantial developer tool I've developed, and something that will allow you to draw maps.