48x48 Map Example
While this may seem long winded to list all locations of tiles for a map, it is also important for how you construct a map. The hashmap is essential, or else this code would be about 2K lines in total alone.

I will continue to play around with making sure we can deal with multiple tilesets used in the same section, since stacking spritesets is vitaly important, and a good example of it done will be important before I handle screen transitions and audio implementation.