48x48 - Additional Info
You can stack elements on top on each other in layers, per the lake code I added that sits upon the grassy plain.

While you do not need to think of this as layers and can draw to whatever you need, it helps if you create foundational layers and stack elements on top and then delete as you desire.

Our max tiles is 50K, meaning for scenes, you shouldn't be running into any problem of overloading the tile entitites.

However, as I continue to look at better ways of making games and using macros, I may add some additional map macros for tiles specifically. Because this may in fact be a global tile max entity limit, which is something we DONT want as the games scale up in size. 50K tiles is the goal per scene, not the goal for the overall project.

I've only set the 50K bound for the sake of you being required to work with scene transitions. I have no requirements on what this needs to be, but fuck, if you are doing 50K tile scenes, you should be smart enough to change the limit of my max tile entities