48x48 Maps - Multiple Tile Map Support
We use another hashmap to manage which tilemap we are referencing, which can be expanded depending on what files you need.

You use indexes as a means of choosing which one you are referencing, and this is based upon the amount of files you have in the hashmap.

This approach allows you to work with as many tilemaps, and also makes sure to blend the textures if needed, which is essential for working with tilemaps