Screen Implementation 01
I've decided to handle screen switches via the appevent place, which then will allow me to put stuff into main.cpp based on which screen we are on.

This is where we want the logic related to screens to go, since this is where we are rendering everything. App event will update the screen via mouse event, and render will update things when we switch screen.

This approach is what will work, given I already have a main screen that has locked us out of our other logic.