CollisionDetection V1 - Player Based Detection
This allows us to manually check for collision based on the players location, and any tile selections we deem as out of bounds.

This is less than ideal compared to tile based detection, and I will modify this far later.

What will need to be done if I want this to work is for the detection of player across tiles, and the passing of keyboard event variables. This can be done if I move my logic to game and find some way of obtaining all the variables associated with it.

Because if I can get these detection from how the player is moving, then I can handle these parameters within the tile themselves. I most likely will take care of half of it when I move a lot of the player logic into AtaraxiaGame, since that will be the ideal place to pull global variables and have global functions help obtain the data needed for colision