CollisionDetection V1.5 - Cross Talk Between Player and mapTile
This is integral for the near completed new collision system.

Instead of relying upon player to detect tiles, we need player's coordinates to interact with each tile it stands upon. This is nearly done, all that is left is to ensure that we can get the tile position, on mapTile when player walks over it, and then create a blocking system in mapTile.

It is about 50% of where it needs to be, since we already have the blocked tiles detected in mapTile, we just need the blocked tiles to be detected here