CollisionDetection V2 - Modular Detection Complete
This code is essential to drawing a map and creating collision bounds.

This makes it to where you can easily set the bounds of the map and any objects that exist inside of it.

This will be essential for 1 thing for later on, an active camera system. Because we don't want the screen to be the limiting factor of where our map is, this will allow users to create a true map for rendering.

We still need to create a map system, since the current approach to mapTile is one that doesn't allow for multiple maps.