MapUpdated - More Tiles
Instead of being limited to what can be drawn inside of the window, we have logic that allows us to render what is OUTSIDE the window.

This is important given our camera tracker changes the definition of what is considered a window, and so hence, we need another renderer for our renderer.

A key component to my SDK, finally incorporated